requirements.txt
supabase>=2.0.0
google-genai>=0.3.0
yfinance>=0.2.0
feedparser>=6.0.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
